rvv: fix vnclip corner case
if shift is 0x3f and vs2[62] is 1, the result is wrong

+++ riscv/insns/vnclip_vx.h
@@ -4,7 +4,7 @@ int64_t int_max = (1 << (P.VU.vsew - 1)) - 1;
 int64_t int_min = -(1 << (P.VU.vsew - 1));
 VI_VVXI_LOOP_NARROW
 ({
-  int64_t result = vs2;
+  int128_t result = vs2;
   // The low log2(2*SEW) bits of the vector or scalar shift amount value are used
   // (e.g., the low 6 bits for a SEW=64-bit to SEW=32-bit narrowing operation).
   uint64_t unsigned_shift_amount = (uint64_t)(rs1 & ((sew * 2) - 1));
