+++ riscv/insns/vssra_vv.h
@@ -3,7 +3,7 @@ VRM xrm = P.VU.get_vround_mode();
 VI_VV_LOOP
 ({
   int sh = vs1 & (sew - 1);
-  int64_t val = vs2;
+  int128_t val = vs2;
 
   INT_ROUNDING(val, xrm, sh);
   vd = val >> sh;
