+++ riscv/insns/vssrl_vv.h
@@ -3,7 +3,7 @@ VRM xrm = P.VU.get_vround_mode();
 VI_VV_ULOOP
 ({
   int sh = vs1 & (sew - 1);
-  uint64_t val = vs2;
+  uint128_t val = vs2;
 
   INT_ROUNDING(val, xrm, sh);
   vd = val >> sh;
